ClickManager Rewrite
Turned CitizenManager to Clickmanager and rewrote code to handle all
clicking from clickManager.
