+++ Assets/Scripts/ClickManager.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 1540d0e51916b244c8b18a7bc89291b9
-timeCreated: 1486603783
+guid: 7eb07715c55a26145897d47060e196e5
+timeCreated: 1486767981
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
